HPCC-14749 Ensure HPCC generates core dump if abort condition sensed.
Add SIGBABRT to handled signals set.

Ensure to create core dump after we caught an abort like signal.

Update after review

+++ ecl/eclagent/eclagent.cpp
@@ -3113,7 +3113,9 @@ void EclAgent::fatalAbort(bool userabort,const char *excepttext)
 #ifdef _WIN32
     TerminateProcess(GetCurrentProcess(), 1);
 #else
-    kill(getpid(), SIGKILL);
+   if (userabort)
+       kill(getpid(), SIGKILL);
+
 #endif
 }
